requirements.txt
# Core dependencies
streamlit
numpy
pandas
matplotlib

# Computer Vision - with system dependencies
opencv-python
Pillow

# Deep Learning & YOLO
--index-url https://download.pytorch.org/whl/cu126
torch
torchvision
torchaudio
ultralytics

# Satellite Image Processing
planetary-computer
pystac-client
pystac
rasterio
pyproj
odc-stac
xarray

# Async Processing
nest_asyncio
aiohttp
async-timeout

# File Management
requests
python-dateutil

# Progress tracking and UI enhancements
tqdm
rich

# Additional utility libraries
scikit-learn
scipy
shapely
geojson
dataclasses
typing-extensions